vcnc-web: widget refactor part 4 - facto a 'TrendWidget' out of 'FrquCacheTrendWidget'
